ppe_compliance_monitoring_system/ppe_compliance_monitoring_train_model/requirements.txt
scikit-learn==1.7.2
scikit-image==0.25.2
numpy==2.3.4
ultralytics==8.3.221
pip-system-certs==5.3
python-dotenv==1.2.1